# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Ecosystem API — CNC Tool Life",
    "description": "Backend ของระบบ CNC Tool Life AI: พยากรณ์อายุดอกกัด (RUL, time series) → ถอดดอก → วัดรอยสึก VB จากภาพใบมีด → ผู้ตรวจยืนยัน → ใบสั่งงาน → retrain\n\nทุก endpoint อยู่ใต้ `/api/v1` · ใช้ปุ่ม **Authorize** ใส่ Bearer token สำหรับ endpoint ที่ต้องล็อกอิน",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/tool-life/fleet": {
      "get": {
        "tags": [
          "Tool Life (RUL)"
        ],
        "summary": "สถานะ RUL ของทุกเครื่อง",
        "operationId": "fleet_api_v1_tool_life_fleet_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/tool-life/machines/{machine}": {
      "get": {
        "tags": [
          "Tool Life (RUL)"
        ],
        "summary": "รายละเอียดเครื่อง + ประวัติรายรัน",
        "operationId": "machine_detail_api_v1_tool_life_machines__machine__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "machine",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Machine"
            }
          },
          {
            "name": "history",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "History"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tool-life/machines/{machine}/speed": {
      "post": {
        "tags": [
          "Tool Life (RUL)"
        ],
        "summary": "ตั้งความเร็วการเล่นซ้ำ",
        "operationId": "set_speed_api_v1_tool_life_machines__machine__speed_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "machine",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Machine"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpeedBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tool-life/machines/{machine}/acknowledge": {
      "post": {
        "tags": [
          "Tool Life (RUL)"
        ],
        "summary": "ตอบสนองคำแนะนำ REPLACE_NOW",
        "operationId": "acknowledge_api_v1_tool_life_machines__machine__acknowledge_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "machine",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Machine"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AckBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tool-life/machines/{machine}/{action}": {
      "post": {
        "tags": [
          "Tool Life (RUL)"
        ],
        "summary": "ควบคุมสตรีม: start / pause / resume / reset",
        "operationId": "control_api_v1_tool_life_machines__machine___action__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "machine",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Machine"
            }
          },
          {
            "name": "action",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Action"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tool-life/model": {
      "get": {
        "tags": [
          "Tool Life (RUL)"
        ],
        "summary": "แบบจำลองที่ใช้งานอยู่ (โหลดจาก MinIO)",
        "operationId": "model_info_api_v1_tool_life_model_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/tool-life/model/versions": {
      "get": {
        "tags": [
          "Tool Life (RUL)"
        ],
        "summary": "ทุกเวอร์ชันใน MinIO bucket models/tool-rul",
        "operationId": "model_versions_api_v1_tool_life_model_versions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/tool-life/model/reload": {
      "post": {
        "tags": [
          "Tool Life (RUL)"
        ],
        "summary": "ดึงแบบจำลองจาก MinIO ใหม่ (admin)",
        "operationId": "model_reload_api_v1_tool_life_model_reload_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ReloadBody"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/tool-life/evaluations": {
      "get": {
        "tags": [
          "Tool Life (RUL)"
        ],
        "summary": "ผลประเมินหลังถอดดอก (เทียบ VB ที่วัดจริง)",
        "operationId": "evaluations_api_v1_tool_life_evaluations_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "detail",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Detail"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AckBody": {
        "properties": {
          "action": {
            "type": "string",
            "title": "Action",
            "description": "continue = ตัดต่อ (override), replace = ถอดดอก"
          }
        },
        "type": "object",
        "required": [
          "action"
        ],
        "title": "AckBody"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReloadBody": {
        "properties": {
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          }
        },
        "type": "object",
        "title": "ReloadBody"
      },
...
      "SpeedBody": {
        "properties": {
          "speed": {
            "type": "number",
            "title": "Speed",
            "description": "หนึ่งใน (1.0, 2.0, 5.0, 10.0, 20.0) (1 = เวลาจริง)"
          }
        },
        "type": "object",
        "required": [
          "speed"
        ],
        "title": "SpeedBody"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}